Batch-Processor/function-http/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery
google-cloud-pubsub
flask
